Revert "6th-lemorning-django-43 추천 리스트 API Refactoring"
