+++ restapi/views.py
@@ -24,6 +24,37 @@ class AccountViewSet(viewsets.ModelViewSet):
     permission_classes = [permissions.IsAuthenticated]
 
 
+class ProductViewSet(viewsets.ModelViewSet):
+    queryset = Product.objects.all()
+    serializer_class = ProductSerializer
+    permission_classes = [permissions.IsAuthenticated]
+
+    def create(self, request, *args, **kwargs):
+        serializer = self.get_serializer(data=request.data)
+        serializer.is_valid(raise_exception=True)
+        self.perform_create(serializer)
+        headers = self.get_success_headers(serializer.data)
+        return Response(serializer.data, status=status.HTTP_201_CREATED, headers=headers)
+
+
+    @action(detail=True, methods=['get'], )
+    def recommand_category(self, request, pk):
+        qs = self.queryset.filter(category_id=pk)
+        serializer = self.get_serializer(qs, many=True)
+        return Response(serializer.data)
+
+    @action(detail=False, methods=['get'], )
+    def weekly_list(self, request):
+        qs = self.queryset.order_by('-created')
+        serializer = self.get_serializer(qs, many=True)
+        response = {
+            'status': status.HTTP_200_OK,
+            'message': "Category List",
+            'response': serializer.data
+        }
+        return Response(response)
+        # return Response(serializer.data)
+
 class CategoryViewSet(viewsets.ModelViewSet):
     queryset = Category.objects.all()
     serializer_class = CategorySerializer
